More safely kill process in ExternalProcessRunner

+++ Core/ExternalProcesses/ExternalProcessReference.cs
@@ -1,4 +1,5 @@
-﻿using System.Diagnostics;
+﻿using System.ComponentModel;
+using System.Diagnostics;
 using System.Threading.Channels;
 using Reductech.EDR.Core.Internal;
 
@@ -34,8 +35,14 @@ public sealed class ExternalProcessReference : IExternalProcessReference
     /// <inheritdoc />
     public void Dispose()
     {
-        Process.Kill(true);
-        Process.Dispose();
+        try
+        {
+            if (!Process.HasExited)
+                Process.Kill(true);
+
+            Process.Dispose();
+        }
+        catch (Win32Exception) { }
     }
 
     /// <inheritdoc />
